Move ProcessedPayload out of MessageEncoder
