+++ src/IO.Ably.Shared/IO.Ably.Shared.projitems
@@ -16,6 +16,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)AblyRest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ApiKey.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MessageEncoders\PayloadCache.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)MessageEncoders\ProcessedPayload.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MessageEncoders\VCDiffEncoder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MessageEncoders\DecodingContext.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MessageEncoders\VcdiffErrorInfo.cs" />
